chore: move file content state logic to filesystem utility and remove inspect module
